+++ test/wasm_worker/proxy_abort_debug.js
@@ -1,0 +1,6 @@
+addToLibrary({
+  proxied_js_function__proxy: 'abort_debug',
+  proxied_js_function: function() {
+  	console.log('In proxied_js_function');
+  }
+});